core: build on musl targets
The libc crate types ptrace requests as c_uint on glibc but c_int on
musl, ioctl requests as c_ulong versus c_int, and msg_controllen as
size_t versus u32. The explicit glibc-shaped casts made a musl build
fail with 19 type errors, so users wanting a static binary for Alpine
had to fall back to a glibc base image. Let the constants carry their
per-target type, cast ioctl requests to libc::Ioctl, and fold the one
helper that named the request type into a PTRACE_CONT-only resume so
no code has to spell the type at all. The statx test constant is
spelled locally because the libc crate only exports it for glibc.

+++ crates/sandlock-core/src/freeze.rs
@@ -119,7 +119,7 @@ fn seize_and_interrupt(tid: i32) -> io::Result<SeizeOutcome> {
     }
 
     let ret = unsafe {
-        libc::ptrace(libc::PTRACE_SEIZE as libc::c_uint, tid, 0, 0)
+        libc::ptrace(libc::PTRACE_SEIZE, tid, 0, 0)
     };
     if ret < 0 {
         let err = io::Error::last_os_error();
@@ -132,7 +132,7 @@ fn seize_and_interrupt(tid: i32) -> io::Result<SeizeOutcome> {
     // before returning so we don't leave the task traced-but-running.
 
     let ret = unsafe {
-        libc::ptrace(libc::PTRACE_INTERRUPT as libc::c_uint, tid, 0, 0)
+        libc::ptrace(libc::PTRACE_INTERRUPT, tid, 0, 0)
     };
     if ret < 0 {
         let err = io::Error::last_os_error();
